api-service/requirements.txt
Flask==2.0.1
requests==2.26.0
kafka-python==2.0.2
Werkzeug>=2.0